backend/requirements.txt
# FastAPI and Server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-decouple>=3.8

# Database
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Data Processing
pandas>=2.1.3
numpy>=1.24.3

# Async Support
aioredis>=2.0.1
asyncio>=3.4.3

# Background Tasks
celery>=5.3.4
redis>=5.0.1

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring and Logging
psutil>=5.9.6
coloredlogs>=15.0.1

# CORS and Static Files
aiofiles>=23.2.1

# WebSocket Support
websockets>=12.0

# Development Tools
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# GPU Monitoring
gputil>=1.4.0